permissions

+++ src/controllers/user.js
@@ -1,48 +1,17 @@
-"use strict";
+"use strict"
 /* -------------------------------------------------------
     NODEJS EXPRESS | CLARUSWAY FullStack Team
 ------------------------------------------------------- */
+// User Controller:
 
-//user Controller
-
-const User = require("../models/user");
-const Token = require("../models/token");
-const jwt = require("jsonwebtoken");
-const passwordEncrypt = require("../helpers/passwordEncrypt");
-const { token } = require("morgan");
-
-const checkUserEmailAndPAssword = function (data) {
-  return data
-  const isEmailValidated = data.email
-    ? /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(data.email)
-    : true;
-
-  if (isEmailValidated) {
-    const isPasswordValidated = data.password
-      ? /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&]).{8,}$/.test(
-          data.password
-        )
-      : true;
-
-    if (isPasswordValidated) {
-      // console.log('Password OK')
-
-      data.password = passwordEncrypt(data.password);
-
-      return data;
-
-      // this._update.password = data.password
-    } else {
-      throw new Error("Password is not validated.");
-    }
-  } else {
-    throw new Error("Email is not validated.");
-  }
-};
+const User = require('../models/user')
+const Token = require('../models/token')
+const passwordEncrypt = require('../helpers/passwordEncrypt')
 
 module.exports = {
-  list: async (req, res) => {
-    /*
+
+    list: async (req, res) => {
+        /*
             #swagger.tags = ["Users"]
             #swagger.summary = "List Users"
             #swagger.description = `
@@ -56,16 +25,21 @@ module.exports = {
             `
         */
 
-    const data = await res.getModelList(User);
-    res.status(200).send({
-      error: false,
-      details: await res.getModelListDetails(User),
-      message: "success",
-      data,
-    });
-  },
-  create: async (req, res) => {
-    /*
+        // Sadece kendi kayıtlarını görebilir:
+        const customFilters = req.user?.isAdmin ? {} : { _id: req.user._id }
+
+        const data = await res.getModelList(User, customFilters)
+
+        res.status(200).send({
+            error: false,
+            details: await res.getModelListDetails(User, customFilters),
+            data
+        })
+
+    },
+
+    create: async (req, res) => {
+        /*
             #swagger.tags = ["Users"]
             #swagger.summary = "Create User"
             #swagger.parameters['body'] = {
@@ -81,44 +55,48 @@ module.exports = {
             }
         */
 
-    const data = await User.create(checkUserEmailAndPAssword(req.body));
-    const tokenData = await Token.create({
-      userId: data._id,
-      token: passwordEncrypt(Date.now() + data._id),
-    });
-    const accessToken = jwt.sign(data.toJSON(), process.env.ACCESS_KEY, {
-      expiresIn: "30m",
-    });
-    const refreshToken = jwt.sign(
-      { _id: data._id, password: data.password },
-      process.env.REFRESH_KEY,
-      {
-        expiresIn: "3d",
-      }
-    );
+        // Yeni kayıtlarda admin/staff = false
+        req.body.isStaff = false
+        req.body.isAdmin = false
 
-    res.status(201).send({
-      error: false,
-      message: "success",
-      token: tokenData.token,
-      bearer: { accessToken, refreshToken },  
-      data,
-    });
-  },
-  read: async (req, res) => {
-    /*
+        const data = await User.create(req.body)
+
+        /* AUTO LOGIN */
+        const tokenData = await Token.create({
+            userId: data._id,
+            token: passwordEncrypt(data._id + Date.now())
+        })
+        /* AUTO LOGIN */
+
+        res.status(201).send({
+            error: false,
+            token: tokenData.token,
+            data
+        })
+
+    },
+
+    read: async (req, res) => {
+        /*
             #swagger.tags = ["Users"]
             #swagger.summary = "Get Single User"
         */
-    const data = await User.findOne({ _id: req.params.id });
-    res.status(200).send({
-      error: false,
-      message: "success",
-      data,
-    });
-  },
-  update: async (req, res) => {
-    /*
+
+        // Sadece kendi kaydını görebilir:
+        const customFilters = req.user?.isAdmin ? { _id: req.params.id } : { _id: req.user._id }
+
+        // const data = await User.findOne({ _id: req.params.id })
+        const data = await User.findOne(customFilters)
+
+        res.status(200).send({
+            error: false,
+            data
+        })
+
+    },
+
+    update: async (req, res) => {
+        /*
             #swagger.tags = ["Users"]
             #swagger.summary = "Update User"
             #swagger.parameters['body'] = {
@@ -134,29 +112,48 @@ module.exports = {
             }
         */
 
-    const data = await User.updateOne(
-      { _id: req.params.id },
-      checkUserEmailAndPAssword(req.body),
-      {
-        runValidators: true,
-      }
-    );
-    res.status(202).send({
-      error: false,
-      message: "success",
-      data,
-      new: await User.findOne({ _id: req.params.id }),
-    });
-  },
-  delete: async (req, res) => {
-    /*
+        // Sadece kendi kaydını güncelleyebilir:
+        const customFilters = req.user?.isAdmin ? { _id: req.params.id } : { _id: req.user._id }
+
+        // Yeni kayıtlarda admin/staff durumunu değiştiremez:
+        if (!req.user?.isAdmin) {
+            delete req.body.isActive
+            delete req.body.isStaff
+            delete req.body.isAdmin
+        }
+        
+        const data = await User.updateOne(customFilters, req.body, { runValidators: true })
+
+        res.status(202).send({
+            error: false,
+            data,
+            new: await User.findOne(customFilters),
+        })
+
+    },
+
+    delete: async (req, res) => {
+        /*
             #swagger.tags = ["Users"]
             #swagger.summary = "Delete User"
         */
-    const data = await User.deleteOne({ _id: req.params.id });
-    res.status(data.deletedCount ? 204 : 404).send({
-      error: !data.deletedCount,
-      data,
-    });
-  },
-};
+
+        // Permission tarafında permissions.isAdmin kontrolü yapıldığı için burda gerek kalmadı.
+
+        if (req.params.id != req.user._id) {
+
+            const data = await User.deleteOne({ _id: req.params.id })
+    
+            res.status(data.deletedCount ? 204 : 404).send({
+                error: !data.deletedCount,
+                data
+            })
+
+        } else {
+            // Admin kendini silemez.
+            res.errorStatusCode = 403
+            throw new Error('You can not remove your account.')
+        }
+    },
+
+}
